fix: small bug fix

+++ backend/app/Http/Controllers/MarcaController.php
@@ -3,88 +3,62 @@
 namespace App\Http\Controllers;
 
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 use App\Models\Marca;
 use App\Models\DescricaoMarca;
+use App\Policies\MarcaPolicy;
 
 class MarcaController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
+    public function __construct()
+    {
+        $this->authorizeResource(Marca::class, 'marca');
+    }
+
     public function index_marca()
     {
+        $this->authorize('viewAny', Marca::class);
+
         return Marca::with('modelos', 'descricao_marca')->get();
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store_marca(Request $request)
-{
-    try {
-        $marca = new Marca();
-
-        $request->validate($marca->rules());
-
-        $novaMarca = Marca::create($request->all());
-
-        return response()->json($novaMarca, 201);
-
-    } catch (\Illuminate\Validation\ValidationException $e) {
-        return response()->json([
-            'message' => 'Já existe uma marca com esse nome!',
-            'errors' => $e->errors()
-        ], 422);
-
-    } catch (\Exception $e) {
-        return response()->json([
-            'message' => 'Erro ao criar a marca.',
-            'error' => $e->getMessage()
-        ], 500);
-    }
-}
-    /**
-     * Display the specified resource.
-     */
-    public function show_marca($id)
     {
-        try {
-            $marca = Marca::with('modelos')->findOrFail($id);
-            return response()->json($marca, 200);
-        } catch (\Exception $e) {
-            return response()->json([
-                'error' => 'Marca não encontrada(ou não existe)',
-                'message' => $e->getMessage()
-            ], 404);
-        }
+        $this->authorize('create', Marca::class);
+
+        $data = $request->validate((new Marca)->rules());
+        // $data ['user_id'] = Auth::id();
+        // $marca = Marca::create($data);
+        $marca = auth()->user()->marcas()->create($data);
+
+        return response()->json($marca, 201);
+    }
+    
+    public function show_marca(Marca $marca)
+    {
+        $this->authorize('view', $marca);
+
+        return response()->json($marca);
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update_marca(Request $request, $id)
+    public function update_marca(Request $request, Marca $marca)
     {
-        $marca = Marca::findOrFail($id);
-
         $this->authorize('update', $marca);
 
-        $request->validate($marca->rules());
+        $data = $request->validate($marca->rules());
 
-        $marca->update($request->all());
+        $marca->update($data);
 
-        return response()->json($marca, 200);
+        return response()->json($marca);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function destroy_marca($id)
-    {
-        $marca = Marca::findOrFail($id);
+    public function destroy_marca(Marca $marca)
+    {    
+        $this->authorize('delete', $marca);
+
         $marca->delete();
 
-        // return response()->json(null, 204);
-        return response()->json(['message' => 'Marca removida com sucesso'], 200);
+        return response()->json(['message' => 'Marca removida com sucesso']);
     }
 }
